refactor: create TrainingPagesComponent for pages under Training section
